Lesson_6/app/requirements.txt
pydantic-settings==2.3.4
sqlmodel==0.0.21
fastapi==0.111.1
uvicorn==0.30.3
joblib==1.4.2
scikit-learn==1.5.1
psycopg-binary==3.2.1
psycopg==3.2.1
pandas==2.2.2
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.0
pydantic==2.8.2
pika==1.3.2
requests==2.32.3
passlib==1.7.4
python-jose==3.3.0
streamlit-authenticator==0.3.3
pytest==8.3.2